CalendarDateEditor edited! CalendarAppointmentListAdapter created! CalendarTime created!
We created for all of these classes xml layouts, menus and string resources.

CalendarDateEditor shows a list with all appointments on a specific date incl. the date itself and the weekday which the date is on.

CalendarAppointmentListAdapter adapts the data from the CalendarDate Object (especially the appointments) to the given list by the layout file.

TODO Add Action for appointments